_btn_pressed_counter: use uint8_t instead of uint16_t.
uint16_t is not required - used values are 0-200.
Avoid using larger data types where not required because MCU is 8bit.

+++ src/inputs.c
@@ -19,7 +19,7 @@ uint8_t _inputs_fall_counter[NO_INPUTS] = {0, };
 uint8_t _btn_debounce_counter = 0;
  // button long press > 2 s
 #define BTN_PRESSED_THRESHOLD (200)
-uint16_t _btn_pressed_counter = 0;
+uint8_t _btn_pressed_counter = 0;
 
 
 static void _inputs_button_debounce_update();
